refactor(solicitacoes): atualiza estrutura de arquivos e remove componente de manutenção; ajusta estilos e renomeia elementos
