+++ byteassist-frontend/src/app/app.component.css
@@ -1,7 +1,8 @@
 .content {
   flex-grow: 1; /* Faz o conteúdo ocupar o espaço restante */
   width: 100%; /* Garante que ocupe toda a largura */
-  min-height: 100vh;
+  height: 100vh;
+  overflow-y: auto;
 }
 
 router-outlet {
