+++ byteassist-frontend/src/app/features/components/menu-sidebar/menu-sidebar.component.html
@@ -90,7 +90,7 @@
           routerLinkActive="active"
         >
           <img src="/images/menu-sidebar/maintenance-f.png" class="menu-icon">
-          <p class="nav-link-text">Manutenção</p>
+          <p class="nav-link-text">Cadastros</p>
         </a>
       </li>
       <!--Employee- END -->
