--- byteassist-frontend/src/app/features/pages-employee/maintenance/maintenance.component.html
@@ -1,3 +1,0 @@
-<section>
-  <h1>Manutenção do funcionario</h1>
-</section>
