--- byteassist-frontend/src/app/features/pages-employee/maintenance/maintenance.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-maintenance',
-  imports: [],
-  templateUrl: './maintenance.component.html',
-  styleUrl: './maintenance.component.css'
-})
-export class MaintenanceComponent {
-
-}
